+++ CSharp/DevVmPowershell/DevVmPsModules/Cmdlets/AddImagesForAllDocumentsModule.cs
@@ -67,8 +67,8 @@ namespace DevVmPsModules.Cmdlets
 				relativityAdminPassword: RelativityAdminPassword,
 				sqlAdminUserName: SqlAdminUserName,
 				sqlAdminPassword: SqlAdminPassword);
-			IImagingHelper imagingHelper = new ImagingHelper(connectionHelper);
 			IRestHelper restHelper = new RestHelper();
+			IImagingHelper imagingHelper = new ImagingHelper(connectionHelper, restHelper, RelativityInstanceName, RelativityAdminUserName, RelativityAdminPassword);
 			IWorkspaceHelper workspaceHelper = new WorkspaceHelper(connectionHelper, restHelper, null, RelativityInstanceName, RelativityAdminUserName, RelativityAdminPassword);
 
 			// Run Imaging Job