+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 -------------------------
 
+- 1/13/2021
+	- REL-507462 - Replace ImagingHelper.cs RSAPI calls with REST implementation.
+
+-------------------------
+
 - 1/12/2021
 	- REL-507447 - Replace WorkspaceHelper.cs RSAPI calls with REST implementation.
 	- REL-507460 - Replace DisclaimerAcceptanceHelper.cs RSAPI calls with REST implementation.